Change snapshots.height field to decimal

+++ db/migrations/20200610125215_create_table_snapshots.js
@@ -7,7 +7,7 @@ exports.up = function (knex) {
       .notNullable()
       .onDelete('CASCADE');
     snapshotsTable.string('plant_uri').notNullable();
-    snapshotsTable.integer('height').notNullable();
+    snapshotsTable.decimal('height').notNullable();
     snapshotsTable.timestamp('created_at').defaultTo(knex.fn.now());
   });
 };
